Fix gettext return type, re-extract Babel strings (#456)
`lazy_gettext` (aliased as `__`) returns `LazyString`, which is not compatible with `str` and causes downstream type errors. Since it's meant to ducktype a string and works in _almost_ all situations (notable exception: `json.dumps`), we re-cast `__` to claim it returns `str`.

Baseframe was also incorrectly exporting `_` and `__` bound to the Baseframe translation domain.